Basic Spring MVC app running with an embedded jetty servlet container bound to port 8180.
Binds the app to the path /emanifest.

One endpoint: /hello

/hello responds with { "message": "Hello, World!" }

That's it.

Basic mvn expectations. `mvn test` runs the tests, `mvn` builds the app
jar.

Run the app from the command line after building like so:

java -jar target/e-manifest-0.0.1.BUILD-SNAPSHOT.jar

http://localhost:8180/emanifest/hello

Add emacs backups, generated artifacts to the gitignore

There are too many things included in the pom. We leave that as an
exercise for the reader.